Add Workaround for Broken Web Requests in the Garmin SDK.
Add a workaround for the brokenness of `makeWebRequest()` in the Garmin
SDK that causes it to 'forget' to add the context to the callback in
some cases. Looks like this can be avoided when the number of in-flight
web requests is limited.
Also fix a bug resulting in refreshes not happening after the first one.